User orders display fix

+++ app/src/main/java/com/mobdeve/s15/animall/OrderModel.kt
@@ -7,6 +7,6 @@ class OrderModel(
     val listingName: String,
     val photosId: String,
     val quantity: Long,
-    val soldPrice: Double,
+    val soldPrice: Long,
     var isConfirmed: Boolean
 )
